requirements.txt
fastapi
uvicorn
requests
python-dotenv
langchain
chromadb
tiktoken
pypdf
python-docx